Test that sidebar link nodes are being scraped
These specs are currently failing, until we make our changes. Unfortunately it
seems there isn't much test coverage in the project already of the parts we're
making changes to, so we need to start with something to reassure us.

Ideally these specs would use test doubles rather than this terrible approach,
but unless anybody suggests a nice way to do that in this particular situation,
we're a bit stuck picturing a mess of nested doubles, stubbing left, right and
centre.