Crawler::Base#hrefs_for
Introduce a method to the Crawler::Base abstract class.

The method takes the 'link nodes' the crawler finds, filters out any that
don't have an href and returns only the existent hrefs.

The existing test coverage wasn't comprehensive enough to catch a typo while
implementing this change across all the crawlers, so we're relying on running a
successful crawl to confirm this works :/