Handle 'case not found' gracefully
We were already rescuing from this and logging a helpful message in the
situation where the crawler finds a 'case detail' that it hasn't previous
processed the corresponding 'case list' for (that is, a case detail that
doesn't have a JSON file to its name).

This change does the same in the situation where the crawler finds an asset
whose case doesn't exist, as described above.