Log unexpected merger case
That is, instead of just blowing up.

What I'm calling an unexpected case is one that hasn't been previously listed
for us by a Case List page. If we follow a link that we identify as a Case, but
that we haven't been told about, there's no JSON file prepared for it in the
_output directory and boom.